requirements.txt
# FastAPI and Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0

# HTTP Requests
requests>=2.31.0
urllib3>=2.0.0
curl_cffi>=0.7.0

# Data Validation
pydantic>=2.0.0

# HTML Parsing
beautifulsoup4>=4.12.0

# Chinese Text Conversion
opencc-python-reimplemented>=0.1.7

# Authentication and Security
PyJWT>=2.8.0
passlib>=1.7.4
argon2-cffi>=23.1.0
google-auth>=2.28.0

lxml==6.0.2
ruff==0.14.14

# Testing
pytest>=8.0
httpx>=0.27
freezegun>=1.2.0